+++ src/App.tsx
@@ -5,7 +5,7 @@ import { BrowserRouter, Route } from "react-router-dom";
 import { Routes } from "react-router-dom";
 import { Page } from "./pages/Page";
 import { Messages } from "./components/messages/Messages";
-import { MyPage } from "./components/MyPage/MyPage";
+import { MyPage } from "./components/myPage/MyPage";
 import { Friends } from "./components/friends/Friends";
 import { Communities } from "./components/communities/Communities";
 
@@ -16,12 +16,11 @@ function App() {
         <Routes>
           <Route path="/" element={<Page />} />
           <Route path="entrance" element={<SignIn />} />
-          <Route path="entrance" element={<SignIn />} />
           <Route path="entrance/gregistration" element={<SignUp />} />
-          <Route path="entrance/messages" element={<Messages />} />
-          <Route path="entrance/mypage" element={<MyPage />} />
-          <Route path="entrance/friends" element={<Friends />} />
-          <Route path="entrance/communities" element={<Communities />} />
+          <Route path="messages" element={<Messages />} />
+          <Route path="mypage" element={<MyPage />} />
+          <Route path="friends" element={<Friends />} />
+          <Route path="communities" element={<Communities />} />
         </Routes>
       </BrowserRouter>
     </>
