+++ src/index.tsx
@@ -1,18 +1,14 @@
-import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
 
-import styled, { createGlobalStyle } from "styled-components";
+import { createGlobalStyle } from "styled-components";
 
 const Global = createGlobalStyle`
 * {
   margin: 0;
   padding: 0;
   box-sizing: border-box;
-
-  
- 
 }`;
 
 const root = ReactDOM.createRoot(
