+++ src/pages/Page.styled.tsx
@@ -1,82 +1,1 @@
 import styled from "styled-components";
-
-export const StyledPageWrapper = styled.div`
-  width: 100%;
-  min-height: 100vh;
-  background: #141414;
-`;
-
-export const StyledPageContainer = styled.div`
-  max-width: 1078px;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 15px;
-  width: 100%;
-  box-sizing: content-box;
-  display: flex;
-  justify-content: space-between;
-`;
-
-export const StyledPageAvatar = styled.div`
-  width: 911px;
-  border-radius: 10px;
-  height: 190px;
-  background: #222222;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  border: 1px solid #373737;
-`;
-
-export const StyledPageFriends = styled.div`
-  width: 345px;
-  border-radius: 10px;
-  height: 550px;
-  background: #222222;
-  margin-top: 15px;
-  padding: 20px;
-  border: 1px solid #373737;
-`;
-
-export const StyledPagePhotoPosts = styled.div`
-  width: 550px;
-  border-radius: 10px;
-`;
-export const StyledPagePhoto = styled.div`
-  width: 550px;
-  border-radius: 10px;
-  height: 350px;
-  background: #222222;
-  margin-top: 15px;
-  padding: 20px;
-  border: 1px solid #373737;
-`;
-
-export const StyledPagePosts = styled.div`
-  margin-top: 15px;
-`;
-
-export const StyledPageWall = styled.div`
-  margin-top: 15px;
-  padding: 20px;
-
-  border-radius: 10px 10px 0 0;
-  width: 550px;
-  background: #222222;
-  height: 64px;
-  border: 1px solid #373737;
-`;
-
-export const StyledPageWall2 = styled.div`
-  padding: 20px;
-  border-radius: 0px 0px 10px 10px;
-  width: 550px;
-  background: #222222;
-  height: 360px;
-  border-bottom: 1px solid #373737;
-  border-left: 1px solid #373737;
-  border-right: 1px solid #373737;
-`;
-
-export const MyPage2 = styled.div``;
-export const MyPage3 = styled.div``;
