+++ src/ui/Text.tsx
@@ -1,5 +1,6 @@
 import styled from "styled-components";
 import React from "react";
+import { MainPage } from "../components/mainPage/MainPage";
 
 export const Text = styled.h1<TextProps>`
   ${({ fs }) => fs && `font-size: ${fs}`};
